pyproject.toml
[project]
name = "tradr"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    # "numpy>=2.0.0",
    # "pandas>=2.1.0",
    # "scikit-learn>=1.3.0",
    # "torch>=2.0.0",
    # "sqlalchemy>=2.0.0",
    # "google-generativeai>=0.3.0",
    "alpaca-py",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0"
]
